requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0

# Agent frameworks (optional)
openai>=1.0.0
anthropic>=0.18.0
langchain>=0.1.0

# Dashboard (optional)
flask>=3.0.0
flask-cors>=4.0.0

# Observability (optional)
prometheus-client>=0.19.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Database (optional)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0
ruff>=0.1.0
